Move file struct
