+++ packages/snaps-utils/coverage.json
@@ -1,6 +1,6 @@
 {
   "branches": 96.44,
-  "functions": 98.63,
-  "lines": 98.74,
-  "statements": 94.49
+  "functions": 98.62,
+  "lines": 98.73,
+  "statements": 94.47
 }
